Code clean up und Bugfix for closing and opening same version window again: property change listener now get registered and unregistered properly

+++ CowGraph/CowGraphModel/src/zbeans/cowgraph/model/GraphElement.java
@@ -44,13 +44,13 @@ public interface GraphElement extends PropertyChangeObservable {
 
     public GraphElementType getType();
 
-    public double getX();
+    public int getX();
 
-    public double getY();
+    public int getY();
 
     public Color getColor();
 
-    public void setX(double x);
+    public void setX(int x);
 
-    public void setY(double y);
+    public void setY(int y);
 }
