+++ CowGraph/CowGraphModel/src/zbeans/cowgraph/model/GraphElementImpl.java
@@ -26,8 +26,8 @@ import zbeans.simple.beans.ObservableBean;
 public class GraphElementImpl extends ObservableBean implements GraphElement {
 
     private final GraphElementType type;
-    private double x;
-    private double y;
+    private int x;
+    private int y;
     private Color color;
 
     public GraphElementImpl(final GraphElementType type) {
@@ -61,7 +61,7 @@ public class GraphElementImpl extends ObservableBean implements GraphElement {
      * @return the value of y
      */
     @Override
-    public double getY() {
+    public int getY() {
         return y;
     }
 
@@ -71,7 +71,7 @@ public class GraphElementImpl extends ObservableBean implements GraphElement {
      * @param y new value of y
      */
     @Override
-    public void setY(double y) {
+    public void setY(int y) {
         double oldY = this.y;
         this.y = y;
         this.firePropertyChange(PROP_Y, oldY, y);
@@ -83,7 +83,7 @@ public class GraphElementImpl extends ObservableBean implements GraphElement {
      * @return the value of x
      */
     @Override
-    public double getX() {
+    public int getX() {
         return x;
     }
 
@@ -93,7 +93,7 @@ public class GraphElementImpl extends ObservableBean implements GraphElement {
      * 
      */
     @Override
-    public void setX(double x) {
+    public void setX(int x) {
         double oldX = this.x;
         this.x = x;
         this.firePropertyChange(PROP_X, oldX, x);
